+++ src/main/java/org/interlink/grouporder/misteram/entity/ShowOrderDTO.java
@@ -4,11 +4,12 @@ import lombok.Getter;
 import lombok.Setter;
 import org.interlink.grouporder.core.entity.MemberOrder;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
 @Setter
 public class ShowOrderDTO {
     List<MemberOrder> members;
-    int fullPrice;
+    BigDecimal fullPrice;
 }
